# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для работы с сервисом расчета генерации электроэнергии ветрогенераторами",
        "title": "Turbines API",
        "contact": {
            "name": "Егор Бляблин"
        },
        "license": {
            "name": "MIT License"
        },
        "version": "0.2.0"
    },
    "host": "localhost:8000",
    "paths": {
        "/api/generation-requests/": {
            "get": {
                "description": "Возвращает список заявок пользователя с фильтрацией по дате и статусу (кроме удаленных и черновиков)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Список заявок с фильтрацией",
                "parameters": [
                    {
                        "type": "string",
                        "name": "formed_at_begin",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "name": "formed_at_end",
                        "in": "path"
                    },
                    {
                        "enum": [
                            "sent",
                            "completed",
                            "rejected"
                        ],
                        "type": "string",
                        "name": "status",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.GenerationRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/generation-requests/draft/": {
            "get": {
                "description": "Возвращает информацию о черновике заявки текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Получение информации о черновике заявки",
                "responses": {
                    "200": {
                        "description": "Информация о черновике",
                        "schema": {
                            "$ref": "#/definitions/ds.GenerationRequest"
                        }
                    },
                    "404": {
                        "description": "Черновик не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет поля черновика заявки текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Обновление черновика заявки",
                "parameters": [
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateGenerationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный черновик",
                        "schema": {
                            "$ref": "#/definitions/ds.GenerationRequest"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Черновик не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет черновик заявки текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Удаление черновика заявки",
                "responses": {
                    "204": {
                        "description": "Черновик удален"
                    },
                    "404": {
                        "description": "Черновик не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/generation-requests/draft/submit/": {
            "put": {
                "description": "Формирует черновик заявки и отправляет на рассмотрение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Отправка черновика заявки",
                "responses": {
                    "200": {
                        "description": "Черновик отправлен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Черновик не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/generation-requests/draft/{turbineId}/": {
            "put": {
                "description": "Обновляет параметры (avg_velocity, alpha) турбины в черновике",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Обновление параметров турбины в черновике",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID турбины",
                        "name": "turbineId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры для обновления",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateTurbineGenerationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Параметры обновлены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Турбина не найдена в черновике",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет турбину в черновик заявки текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Добавление турбины в черновик",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID турбины",
                        "name": "turbineId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Турбина добавлена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "304": {
                        "description": "Турбина уже в черновике",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет турбину из черновика заявки текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки расчета выработки"
                ],
                "summary": "Удаление турбины из черновика",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID турбины",
                        "name": "turbineId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Турбина удалена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Турбина или черновик не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.GenerationRequest": {
            "type": "object",
            "properties": {
                "calculated_generation_sum": {
                    "type": "integer"
                },
                "closed_at": {
                    "type": "string"
                },
                "closed_by": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "formed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "period_days": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "turbine_generation_requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.TurbineGenerationRequest"
                    }
                },
                "turbine_generation_requests_count": {
                    "type": "integer"
                }
            }
        },
        "ds.Turbine": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "power": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ds.TurbineGenerationRequest": {
            "type": "object",
            "properties": {
                "alpha": {
                    "type": "number"
                },
                "avgVelocity": {
                    "type": "number"
                },
                "calculatedGeneration": {
                    "type": "integer",
                    "format": "int64"
                },
                "generationRequestID": {
                    "type": "integer"
                },
                "turbine": {
                    "$ref": "#/definitions/ds.Turbine"
                },
                "turbineID": {
                    "type": "integer"
                }
            }
        },
        "ds.UpdateGenerationRequest": {
            "type": "object",
            "properties": {
                "period_days": {
                    "type": "integer"
                }
            }
        },
...
        "ds.UpdateTurbineGenerationRequest": {
            "type": "object",
            "properties": {
                "alpha": {
                    "type": "number"
                },
                "avg_velocity": {
                    "type": "number"
                }
            }
        },
...
    }
}